// types/Notification.data.ts
export enum NotificationTitles {
    STAKE_CLAIM = 'Claiming stake',
    STAKE_CREATE = 'Creating stake',
    STAKE_RESTAKE = 'Restaking item',
    STAKE_INIT = 'Staking initialization',
    STAKE_CLAIM_MULTIPLE = 'Claiming stakes',
    STAKE_MERGE = 'Merging stakes',

    AH_LIST = 'Listing item',
    AH_CANCEL_LIST = 'Cancel Listing',
    AH_ESCROW_WITHDRAWAL = 'Escrow withdrawal',
    AH_PURCHASE_CONFIRMATION = 'Purchase confirmation',

    DAO_SUBMIT_VOTE = 'Submit vote',

    SWAP = 'Swapping tokens',

    TEXT_COPIED = 'Text has been successfully copied',

    GENERAL_ERROR = 'An error occured while executing action',

    PARAGONS_STAMP = 'Stamping paragon',
    PARAGONS_FUSE = 'Fusing paragons',
    PARAGONS_BREED = 'Breeding paragons',
    PARAGONS_CLAIM_BREEDING = 'Claiming paragons',
    PARAGONS_CLAIM_FUSING = 'Claiming paragon',
    PARAGONS_CLAIM_MULTIPLE = 'Claiming paragons',

    TREASURY_CLAIM_CASHBACK = 'Claiming cashback',

    RPC_CHANGED = 'RPC switch',
}